fix: rename backend/ to lib/ to avoid Decky CLI build conflict
Decky CLI treats a backend/ directory as a custom backend container
and expects a Dockerfile or entrypoint.sh. Rename to lib/ since this
is just a Python package of mixin modules.

Updates all imports, .importlinter, mise.toml deploy paths, and docs.
