+++ .gitignore
@@ -35,3 +35,4 @@ PLAN.md
 
 # Claude Code
 .claude/
+.venv/
