--- backend/__init__.py
@@ -1,13 +1,0 @@
-from backend.state import StateMixin
-from backend.romm_client import RommClientMixin
-from backend.sgdb import SgdbMixin
-from backend.steam_config import SteamConfigMixin
-from backend.firmware import FirmwareMixin, BIOS_DEST_MAP
-from backend.metadata import MetadataMixin
-from backend.downloads import DownloadMixin
-from backend.sync import SyncMixin
-
-__all__ = [
-    "StateMixin", "RommClientMixin", "SgdbMixin", "SteamConfigMixin",
-    "FirmwareMixin", "BIOS_DEST_MAP", "MetadataMixin", "DownloadMixin", "SyncMixin",
-]
