+++ lib/firmware.py
